Reformat python code using black
I do not like the idea of a tool changing my code without my explicit
approval of the specific change. My recollection is that it is
non-trivial to do this *properly* with a pre-commit hook, so for it will
be a manual process.

For the unaware, "black" is a utility for reformatting python code:

https://pypi.org/project/black/